// codegen/Faucet.ts
export const FaucetContractArtifact = loadContractArtifact(FaucetContractArtifactJson as NoirCompiledContract);



/**
 * Type-safe interface for contract Faucet;
 */
export class FaucetContract extends ContractBase {
  
  protected constructor(
    address: AztecAddress,
    wallet: Wallet,
  ) {
    super(address, FaucetContractArtifact, wallet);
  }
  

  
  /**
   * Creates a contract instance.
   * @param address - The deployed contract's address.
   * @param wallet - The wallet to use when interacting with the contract.
   * @returns A new Contract instance.
   */
  public static at(
    address: AztecAddress,
    wallet: Wallet,
  ): FaucetContract {
    return Contract.at(address, FaucetContract.artifact, wallet) as FaucetContract;
  }

  
  /**
   * Creates a tx to deploy a new instance of this contract.
   */
  public static deploy(wallet: Wallet, token: AztecAddressLike, admin: AztecAddressLike, drip_amount: (bigint | number), cooldown_seconds: (bigint | number)) {
    return new DeployMethod<FaucetContract>(PublicKeys.default(), wallet, FaucetContractArtifact, (instance, wallet) => FaucetContract.at(instance.address, wallet), Array.from(arguments).slice(1));
  }

  /**
   * Creates a tx to deploy a new instance of this contract using the specified public keys hash to derive the address.
   */
  public static deployWithPublicKeys(publicKeys: PublicKeys, wallet: Wallet, token: AztecAddressLike, admin: AztecAddressLike, drip_amount: (bigint | number), cooldown_seconds: (bigint | number)) {
    return new DeployMethod<FaucetContract>(publicKeys, wallet, FaucetContractArtifact, (instance, wallet) => FaucetContract.at(instance.address, wallet), Array.from(arguments).slice(2));
  }

  /**
   * Creates a tx to deploy a new instance of this contract using the specified constructor method.
   */
  public static deployWithOpts<M extends keyof FaucetContract['methods']>(
    opts: { publicKeys?: PublicKeys; method?: M; wallet: Wallet },
    ...args: Parameters<FaucetContract['methods'][M]>
  ) {
    return new DeployMethod<FaucetContract>(
      opts.publicKeys ?? PublicKeys.default(),
      opts.wallet,
      FaucetContractArtifact,
      (instance, wallet) => FaucetContract.at(instance.address, wallet),
      Array.from(arguments).slice(1),
      opts.method ?? 'constructor',
    );
  }
  

  
  /**
   * Returns this contract's artifact.
   */
  public static get artifact(): ContractArtifact {
    return FaucetContractArtifact;
  }

  /**
   * Returns this contract's artifact with public bytecode.
   */
  public static get artifactForPublic(): ContractArtifact {
    return loadContractArtifactForPublic(FaucetContractArtifactJson as NoirCompiledContract);
  }
  

  public static get storage(): ContractStorageLayout<'config' | 'last_drip'> {
      return {
        config: {
      slot: new Fr(1n),
    },
last_drip: {
      slot: new Fr(6n),
    }
      } as ContractStorageLayout<'config' | 'last_drip'>;
    }
    

  /** Type-safe wrappers for the public methods exposed by the contract. */
  public declare methods: {
    
    /** admin_drip(recipient: struct, amount: integer) */
    admin_drip: ((recipient: AztecAddressLike, amount: (bigint | number)) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** constructor(token: struct, admin: struct, drip_amount: integer, cooldown_seconds: integer) */
    constructor: ((token: AztecAddressLike, admin: AztecAddressLike, drip_amount: (bigint | number), cooldown_seconds: (bigint | number)) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** drip(recipient: struct) */
    drip: ((recipient: AztecAddressLike) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** get_config() */
    get_config: (() => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** get_last_drip(recipient: struct) */
    get_last_drip: ((recipient: AztecAddressLike) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** offchain_receive(messages: struct) */
    offchain_receive: ((messages: { ciphertext: FieldLike[], recipient: AztecAddressLike, tx_hash: OptionLike<FieldLike>, anchor_block_timestamp: (bigint | number) }[]) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** public_dispatch(selector: field) */
    public_dispatch: ((selector: FieldLike) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** sync_state(scope: struct) */
    sync_state: ((scope: AztecAddressLike) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;
  };

  
}