// codegen/FPCMultiAsset.ts
export const FPCMultiAssetContractArtifact = loadContractArtifact(FPCMultiAssetContractArtifactJson as NoirCompiledContract);



/**
 * Type-safe interface for contract FPCMultiAsset;
 */
export class FPCMultiAssetContract extends ContractBase {
  
  protected constructor(
    address: AztecAddress,
    wallet: Wallet,
  ) {
    super(address, FPCMultiAssetContractArtifact, wallet);
  }
  

  
  /**
   * Creates a contract instance.
   * @param address - The deployed contract's address.
   * @param wallet - The wallet to use when interacting with the contract.
   * @returns A new Contract instance.
   */
  public static at(
    address: AztecAddress,
    wallet: Wallet,
  ): FPCMultiAssetContract {
    return Contract.at(address, FPCMultiAssetContract.artifact, wallet) as FPCMultiAssetContract;
  }

  
  /**
   * Creates a tx to deploy a new instance of this contract.
   */
  public static deploy(wallet: Wallet, operator: AztecAddressLike, operator_pubkey_x: FieldLike, operator_pubkey_y: FieldLike) {
    return new DeployMethod<FPCMultiAssetContract>(PublicKeys.default(), wallet, FPCMultiAssetContractArtifact, (instance, wallet) => FPCMultiAssetContract.at(instance.address, wallet), Array.from(arguments).slice(1));
  }

  /**
   * Creates a tx to deploy a new instance of this contract using the specified public keys hash to derive the address.
   */
  public static deployWithPublicKeys(publicKeys: PublicKeys, wallet: Wallet, operator: AztecAddressLike, operator_pubkey_x: FieldLike, operator_pubkey_y: FieldLike) {
    return new DeployMethod<FPCMultiAssetContract>(publicKeys, wallet, FPCMultiAssetContractArtifact, (instance, wallet) => FPCMultiAssetContract.at(instance.address, wallet), Array.from(arguments).slice(2));
  }

  /**
   * Creates a tx to deploy a new instance of this contract using the specified constructor method.
   */
  public static deployWithOpts<M extends keyof FPCMultiAssetContract['methods']>(
    opts: { publicKeys?: PublicKeys; method?: M; wallet: Wallet },
    ...args: Parameters<FPCMultiAssetContract['methods'][M]>
  ) {
    return new DeployMethod<FPCMultiAssetContract>(
      opts.publicKeys ?? PublicKeys.default(),
      opts.wallet,
      FPCMultiAssetContractArtifact,
      (instance, wallet) => FPCMultiAssetContract.at(instance.address, wallet),
      Array.from(arguments).slice(1),
      opts.method ?? 'constructor',
    );
  }
  

  
  /**
   * Returns this contract's artifact.
   */
  public static get artifact(): ContractArtifact {
    return FPCMultiAssetContractArtifact;
  }

  /**
   * Returns this contract's artifact with public bytecode.
   */
  public static get artifactForPublic(): ContractArtifact {
    return loadContractArtifactForPublic(FPCMultiAssetContractArtifactJson as NoirCompiledContract);
  }
  

  public static get storage(): ContractStorageLayout<'config'> {
      return {
        config: {
      slot: new Fr(1n),
    }
      } as ContractStorageLayout<'config'>;
    }
    

  /** Type-safe wrappers for the public methods exposed by the contract. */
  public declare methods: {
    
    /** cold_start_entrypoint(user: struct, accepted_asset: struct, bridge: struct, claim_amount: integer, claim_secret: field, claim_secret_hash: field, message_leaf_index: field, fj_fee_amount: integer, aa_payment_amount: integer, valid_until: integer, quote_sig: array) */
    cold_start_entrypoint: ((user: AztecAddressLike, accepted_asset: AztecAddressLike, bridge: AztecAddressLike, claim_amount: (bigint | number), claim_secret: FieldLike, claim_secret_hash: FieldLike, message_leaf_index: FieldLike, fj_fee_amount: (bigint | number), aa_payment_amount: (bigint | number), valid_until: (bigint | number), quote_sig: (bigint | number)[]) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** constructor(operator: struct, operator_pubkey_x: field, operator_pubkey_y: field) */
    constructor: ((operator: AztecAddressLike, operator_pubkey_x: FieldLike, operator_pubkey_y: FieldLike) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** fee_entrypoint(accepted_asset: struct, authwit_nonce: field, fj_fee_amount: integer, aa_payment_amount: integer, valid_until: integer, quote_sig: array) */
    fee_entrypoint: ((accepted_asset: AztecAddressLike, authwit_nonce: FieldLike, fj_fee_amount: (bigint | number), aa_payment_amount: (bigint | number), valid_until: (bigint | number), quote_sig: (bigint | number)[]) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** offchain_receive(messages: struct) */
    offchain_receive: ((messages: { ciphertext: FieldLike[], recipient: AztecAddressLike, tx_hash: OptionLike<FieldLike>, anchor_block_timestamp: (bigint | number) }[]) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** public_dispatch(selector: field) */
    public_dispatch: ((selector: FieldLike) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;

    /** sync_state(scope: struct) */
    sync_state: ((scope: AztecAddressLike) => ContractFunctionInteraction) & Pick<ContractMethod, 'selector'>;
  };

  
}